refactor: consolidate test files — merge _cov2 pairs, standardize names
- Merge 7 redundant _cov/_cov2 pairs into base files (unique tests preserved)
- Rename 23 _cov2-only files to standard names (drop suffix)
- Delete 7 _cov2 files after merge
- 5,452 tests pass, zero regressions

Closes #571
