+++ tests/test_azure_cis.py
