+++ tests/test_cli_analysis.py
