+++ tests/test_cli_history.py
