+++ tests/test_cli_inventory.py
