+++ tests/test_cli_registry.py
