+++ tests/test_cli_server.py
