+++ tests/test_cloud_azure.py
