+++ tests/test_watch_cov.py
@@ -13,6 +13,8 @@ from agent_bom.watch import (
     ConsoleAlertSink,
     FileAlertSink,
     WebhookAlertSink,
+    discover_config_dirs,
+    discover_config_paths,
 )
 
 
@@ -154,3 +156,84 @@ class TestConfigChangeHandler:
         handler._process_diff(diff, "/config")
         call_args = mock_sink.send.call_args[0][0]
         assert call_args.severity == "medium"
+
+
+# ── Unique tests from cov2 ──────────────────────────────────────────────────
+
+
+class TestAlertTimestamps:
+    def test_auto_timestamp(self):
+        a = Alert(alert_type="test", severity="info", summary="hello")
+        assert a.timestamp != ""
+        assert "T" in a.timestamp
+
+    def test_custom_timestamp(self):
+        a = Alert(timestamp="2025-01-01T00:00:00Z")
+        assert a.timestamp == "2025-01-01T00:00:00Z"
+
+
+class TestConsoleAlertSinkExtraSeverities:
+    def test_unknown_severity(self):
+        sink = ConsoleAlertSink()
+        alert = Alert(severity="unknown", summary="test unknown")
+        sink.send(alert)
+
+
+class TestWebhookAlertSink4xx:
+    def test_4xx_no_retry(self):
+        mock_resp = MagicMock()
+        mock_resp.status_code = 403
+
+        with patch("httpx.post", return_value=mock_resp) as mock_post:
+            sink = WebhookAlertSink("https://hooks.example.com/test", retries=2)
+            alert = Alert(severity="info", summary="forbidden")
+            sink.send(alert)
+            assert mock_post.call_count == 1
+
+
+class TestHandlerDebounceWithScanAndAlert:
+    def test_debounce_skips_second_call(self):
+        sinks = [MagicMock()]
+        handler = ConfigChangeHandler(sinks, debounce_seconds=10.0)
+        with patch.object(handler, "_scan_and_alert") as mock_scan:
+            handler.on_modified("/test/config.json")
+            assert mock_scan.call_count == 1
+            handler.on_modified("/test/config.json")
+            assert mock_scan.call_count == 1
+
+
+class TestHandlerProcessDiffCritical:
+    def test_critical_severity(self):
+        sink = MagicMock()
+        handler = ConfigChangeHandler([sink])
+        diff = {
+            "summary": {"new_findings": 1, "new_packages": 0},
+            "new": [{"severity": "critical"}],
+        }
+        handler._process_diff(diff, "/test")
+        alert = sink.send.call_args[0][0]
+        assert alert.severity == "critical"
+
+
+class TestHandlerProcessDiffNewPackages:
+    def test_new_packages_config_changed(self):
+        sink = MagicMock()
+        handler = ConfigChangeHandler([sink])
+        diff = {
+            "summary": {"new_findings": 0, "new_packages": 5},
+            "new": [],
+        }
+        handler._process_diff(diff, "/test")
+        sink.send.assert_called_once()
+        alert = sink.send.call_args[0][0]
+        assert alert.alert_type == "config_changed"
+
+
+class TestDiscoverHelpers:
+    def test_discover_config_paths(self):
+        paths = discover_config_paths()
+        assert isinstance(paths, list)
+
+    def test_discover_config_dirs(self):
+        dirs = discover_config_dirs()
+        assert isinstance(dirs, list)
